// js_modules/dagster-ui/packages/ui-core/src/assets/auto-materialization/types/AutomaterializationRoot.types.ts
// Generated GraphQL types, do not edit manually.

import * as Types from '../../../graphql/types';

export type AssetDameonTicksQueryVariables = Types.Exact<{
  dayRange?: Types.InputMaybe<Types.Scalars['Int']>;
  dayOffset?: Types.InputMaybe<Types.Scalars['Int']>;
  statuses?: Types.InputMaybe<Array<Types.InstigationTickStatus> | Types.InstigationTickStatus>;
  limit?: Types.InputMaybe<Types.Scalars['Int']>;
  cursor?: Types.InputMaybe<Types.Scalars['String']>;
}>;

export type AssetDameonTicksQuery = {
  __typename: 'Query';
  autoMaterializeTicks: Array<{
    __typename: 'InstigationTick';
    id: string;
    timestamp: number;
    endTimestamp: number | null;
    status: Types.InstigationTickStatus;
    instigationType: Types.InstigationType;
    requestedAssetMaterializationCount: number;
    autoMaterializeAssetEvaluationId: number | null;
    error: {
      __typename: 'PythonError';
      message: string;
      stack: Array<string>;
      errorChain: Array<{
        __typename: 'ErrorChainLink';
        isExplicitLink: boolean;
        error: {__typename: 'PythonError'; message: string; stack: Array<string>};
      }>;
    } | null;
    requestedAssetKeys: Array<{__typename: 'AssetKey'; path: Array<string>}>;
  }>;
};

export type AssetDaemonTickFragment = {
  __typename: 'InstigationTick';
  id: string;
  timestamp: number;
  endTimestamp: number | null;
  status: Types.InstigationTickStatus;
  instigationType: Types.InstigationType;
  requestedAssetMaterializationCount: number;
  autoMaterializeAssetEvaluationId: number | null;
  error: {
    __typename: 'PythonError';
    message: string;
    stack: Array<string>;
    errorChain: Array<{
      __typename: 'ErrorChainLink';
      isExplicitLink: boolean;
      error: {__typename: 'PythonError'; message: string; stack: Array<string>};
    }>;
  } | null;
  requestedAssetKeys: Array<{__typename: 'AssetKey'; path: Array<string>}>;
};